automation/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.13.3
bitarray==2.9.3
bitstring==4.2.3
bs4==0.0.2
cachetools==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
discord.py==2.4.0
ecdsa==0.19.0
esptool==4.8.1
flatbuffers==25.1.21
fonttools==4.55.5
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
intelhex==2.3.0
jax==0.5.0
jaxlib==0.5.0
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.3.1
matplotlib==3.10.0
mediapipe==0.10.20
ml_dtypes==0.5.1
MouseInfo==0.1.3
multidict==6.1.0
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
playwright==1.50.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
PyYAML==6.0.2
reedsolo==1.7.0
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.15.1
sentencepiece==0.2.0
six==1.17.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
tqdm==4.67.1
tweepy==4.14.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
yarl==1.18.3
